agentic-ui-v0/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
azure-ai-agents==1.2.0b3
azure-ai-inference==1.0.0b9
azure-ai-projects==1.1.0b3
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
jsonref==1.1.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
opentelemetry-api==1.36.0
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
